Update patch set 3
Patch Set 3:

> Thanks for the contribution Alexander. I'll take a look as soon as
 > I can.
Thank you, please take your time!

 > I did want to comment about the config option of channel and
 > username. This is nice as you can setup slack once and have per
 > project overrides if you want different projects to post to
 > different channels using the same webhook. I see you no longer have
 > that change in the current patchset but wanted to explain why those
 > options were there.
Yes, i noticed that and removed it in newer patch sets.

There are also more options to override, which we might want to have in future, for example the icon/avatar of the bot.

but that would be done in another CL.

Patch-set: 3