Update patch set 2
Patch Set 2:

As far as I can see this change is the same on all the branches.

Since there are already merges done from stable-2.9 upwards, presumably already resolving any conflicts, submitting this on stable-2.9 and then performing a merge up to the other branches will only result in this change being merged up.

By merging it's also clearer (by doing git log or gitk) that everything on stable-x is included on stable-y, etc.

It's your plugin though, so it's up to you how you want to do it :)

Patch-set: 2