Update patch set 2
Patch Set 2:

I am so missing some fundamental git thing here that I've not encountered. The conflict for this was around 2.12 to 2.13 plus other branch wide api chances. If I put this on stable-2.9 would I not have to deal with the unrelated changes to the api signatures in the merge to newer branches?

Patch-set: 2