Update patch set 1
Patch Set 1:

You are right the documentation is missing for this change. Besides, what you see here is the squashed commit of one of the changes I did (more in the next section). In my local repository I have the original five commits, but the drawback is not all of those commits give green tests. This is the main reason I squashed those to this larger commit. I can push the single commits instead if you prefer. 

Regarding the changes I did, here is a quick overview:
#1 (in this change, 5 commits): The use case is you have a repository with multiple branches, where projects are being developed on. If you now assume a single slack channel for every project this change will allow you to create a plugin configuration for every branch and by this publish events for different branches to the corresponding project channel in slack.
#2 (not pushed yet, 1 commit, meanwhile, more a refactoring than a feature) For a configuration you can specify which event shall be published to slack and which not. This is similar to what Nathan Wall did, but instead of introducing a boolean for every event, I use a white list containing the name of the event.

Here is an excerpt of the updated README.md to give you an idea of the change:

<snip>
Create the following config block

    vi project.config

    [plugin "slack-integration"]
        enabled = true
        webhookurl = https://<web-hook-url>
        channel = general
        username = gerrit
        ignore = "^WIP.*"
        events = patchSetCreated, commentAdded, changeMerged

In case you have multiple branches for a project, which shall post events to
different Slack channels, a plugin specific configuration file is used. Instead
storing your configuration to _project.config_ you create a _slack-integration.config_
and put your configuration in there.

    vi slack-integration.config

    [branch "master"]
        enabled = true
        webhookurl = https://<web-hook-url>
        channel = general
        username = gerrit
        ignore = "^WIP.*"
        events = patchSetCreated, commentAdded, changeMerged

Commit and push changes

    git commit -a
    git push origin meta/config:meta/config

[...]

   events - String (optional to keep legacy behavior)
        A filter that, if defined, allows to define a white list of events
        that are published by the plugin. Valid entries are:
        changeMerged
            Publishes a message in case a change is merged into the repository
        commentAdded
            Publishes a message in case a comment is added to a change
        patchSetCreated
            Publishes a message in case a new patch set is created at Gerrit
        reviewed
            Publishes a messages in case a code review (+2) is done
        verified
            Publishes a message in case a change is verified (+1)
</snip>

For #1 I need then to know, if you like to review the single commits although not every single one has working tests. For #2 I need to know if you like the idea of having a white list, in which case I will rebase my changes for #1 on the current master.

Patch-set: 1
Reviewer: Soeren Kewenig <1031294@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0