Update patch set 1
Patch Set 1:

Hi Soeren,

First off, sorry this has take so long for me to get back to. I've been traveling a 
lot and been super busy with work so far this year. Thanks for taking the time 
to reply. 

 > Regarding the changes I did, here is a quick overview:
 > #1 (in this change, 5 commits): The use case is you have a
 > repository with multiple branches, where projects are being
 > developed on. If you now assume a single slack channel for every
 > project this change will allow you to create a plugin configuration
 > for every branch and by this publish events for different branches
 > to the corresponding project channel in slack.

I've not run into the need to have specific branches publish to different slack 
channels and I honestly feel that mixing the configuration of the plugin with
the actual project code isn't the right direction for a general plugin. The
Buit-in Gerrit configuration model works very well in that you can specify a
base configuration and only alter what you need per project using the built-in
configuration inheritance support. I am not sure I'd want to build in support 
for a plugin specific configuration file. I'd much rather see configuration contained to
the refs/meta/config/project.config.

 > #2 (not pushed yet, 1 commit, meanwhile, more a refactoring than a
 > feature) For a configuration you can specify which event shall be
 > published to slack and which not. This is similar to what Nathan
 > Wall did, but instead of introducing a boolean for every event, I
 > use a white list containing the name of the event.

The default state for all supported events is to publish. While I like the idea
of a single line denoting which events to publish or not publish, a white list of 
events goes against the default state. I'd be interested in seeing a single review
that proposed a black list of events to not publish as that would be more in line
with how the current plugin works. I'd be somewhat hesitant though as 
this is a change that would break existing uses of the plugin.

Patch-set: 1
Reviewer: Gerrit User 1015822 <1015822@173816e5-2b9a-37c3-8a2e-48639d4f1153>