Update patch set 1
Patch Set 1:

I uploaded the smaller changes, but they are not yet rebased against the newest changes from yesterday. I will do this if we agree on the relevance of my changes.

Regarding the use case: The company I am working for has a product that is being maintained and developed since more than 10 years. The code base has more than 10k source files. On this product multiple projects are developed in feature branches, where one branch can last for up to 3 years.
The idea is that every project team has a separate slack channel to exchange information. In those channels also jenkins and gerrit post their information, but only those relevant for the project.

Now if you want to support this use case, it is to my understanding not possible to use the project config file, because you can have only a single section for the plugin. So the questions was: how to provide this configuration per branch? The recommend approach I found is to use a separate config file for the plugin [1], which contains a list of configurations - a config block as you know per branch. The downside is, that you loose a bit of convenience, because it is not possiblt to edit a plugin specific configuration file directly in gerrits project options. To not break compatibility my implementation does currently support both cases.
I am a bit surprised this is an edge case, because larger code bases with multiple branches are not seldom. But maybe those companies are using one of the selfhosted slack clones instead of slack itself.

I also try to push the change #2, which will show you that the downwards compatibility is given, because it the whitelist is only used in case the whitelist is present in the configuration. Othewise the default behavior is to publish, again to not break the existing behavior.

[1] https://gerrit-review.googlesource.com/Documentation/dev-plugins.html#project-specific-configuration

Patch-set: 1
Reviewer: Gerrit User 1031294 <1031294@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0