Update patch set 1
Patch Set 1:

Actually, let me respond inline on this one.
 
 > Regarding the use case: The company I am working for has a product
 > that is being maintained and developed since more than 10 years.
 > The code base has more than 10k source files. On this product
 > multiple projects are developed in feature branches, where one
 > branch can last for up to 3 years.
 > The idea is that every project team has a separate slack channel to
 > exchange information. In those channels also jenkins and gerrit
 > post their information, but only those relevant for the project.

Thanks for the explanation of what you were solving for.

 > Now if you want to support this use case, it is to my understanding
 > not possible to use the project config file, because you can have
 > only a single section for the plugin. So the questions was: how to
 > provide this configuration per branch? The recommend approach I
 > found is to use a separate config file for the plugin [1], which
 > contains a list of configurations - a config block as you know per
 > branch. The downside is, that you loose a bit of convenience,
 > because it is not possiblt to edit a plugin specific configuration
 > file directly in gerrits project options. To not break
 > compatibility my implementation does currently support both cases.
 > I am a bit surprised this is an edge case, because larger code
 > bases with multiple branches are not seldom. But maybe those
 > companies are using one of the selfhosted slack clones instead of
 > slack itself.

I am coming around to the idea of this in general if I am understanding what you are looking to solve for correctly. I came up with the following requirements.

1. Allow for branch specific configuration using the plugin specific config file.
2. Support the full inheritance of config options meaning that I can set options at the base project config, set other options at the inheriting project level and further set branch specific options using the plugin specific config file.

I would like to see this included. For this I'd want to see one review off current master that implements just those requirements and that adheres to the same commit message style that David referenced as well as whitespace style that the rest of the codebase uses. Initially I'd like to stick with the exact config options that are currently supported for this review. Meaning that the config block used in the project config will be supported in the plugin specific config file with no additions or other alterations.  

I do think we can work on minimizing the code changes needed for this as compared to your original implementation but getting a cohesive easy to digest review for this is step one. 

 > I also try to push the change #2, which will show you that the
 > downwards compatibility is given, because it the whitelist is only
 > used in case the whitelist is present in the configuration.
 > Othewise the default behavior is to publish, again to not break the
 > existing behavior.

Let's work on the first review to start with. Once that is landed we can take a look at optional ways in which to enable/disable specific publish events.

If this sounds like an acceptable way forward let me know. Please abandon any outstanding reviews and resubmit based this discussion.

Thanks!

Patch-set: 1
Reviewer: Gerrit User 1015822 <1015822@173816e5-2b9a-37c3-8a2e-48639d4f1153>