Update patch set 2
Patch Set 2:

I am new to Bazel but have been interested in converting this plugin from a maven build to bazel to better map to the rest of Gerrit. So thanks for this! 

Should I expect "bazel build" to result in building the plugin artifact or am I doing something wrong.

Patch-set: 2
CC: Matthew Montgomery <1015822@173816e5-2b9a-37c3-8a2e-48639d4f1153>