Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Speaking of Maven, do you all publish the SNAPSHOT plugin API anywhere? I'd like to be able to update the POM  and have the gerrit-forge CI job for master no break due to a missing dependency.

I'm not sure what CI is doing for Maven built plugins, but IIUC for bazel it always build in-tree so this isn't an issue.

As for snapshots, yes but no not really.  The snapshot builds can be deployed to Maven Central, but it's not something that we regularly do.  If you need one, just let me know and I'll do it.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>