Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Shouldn't this change be on stable-2.14 which is the version in which we started to use google-java-format?

It was done on stable-2.14 with [1].  Note that on this project the branches are being maintained separately; i.e. the stable branches are not merged up to master and all commits are cherry-picked or redone per branch.

[1] https://gerrit-review.googlesource.com/c/plugins/slack-integration/+/165630

Patch-set: 1
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>