Update patch set 1
Patch Set 1:

It's not as easy as I thought it would be to reorder these commits.  If the AllProjectsName is injected in the constructor, it doesn't make sense to leave the ProjectconfigFactory outside of the constructor.

I would suggest we either squash the changes, or just merge them both now and handle the cleanup of tests later.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>