Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> When I build locally with Maven I am not seeing any issues. Are you guy still seeing problems there?

The only issue left that I'm aware of started before this change. It is about failing this plugin's tests when building or running them in gerrit tree using bazel. See the aforementioned NoClassDefFoundError sample.

Since that issue was there before this change, and given that we are now done positively reviewing this change, we can then submit it. That existing issue is to be fixed eventually, then.

Patch-set: 2
Reviewer: Gerrit User 1012732 <1012732@173816e5-2b9a-37c3-8a2e-48639d4f1153>