Update patch set 1
Patch Set 1:

Why remove the Maven build, if maintainer of the project prefer it? My point of broken CI build was only because of Snapshot plugin API that we do not publish currently. Of course the plugin can be still built with Maven on stable branches, where the plugin API is published on Maven Central for sure.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>