Add config option for ignoring comments from specific author(s).
The change introduces a new configuration option, ignore-comment-author,
which allows filtering the comment-added events based on the author.
The motivation behind the change is to be able to ignore automated build
messages from a CI/CD system integrated with gerrit (e.g. Jenkins),
thus keeping the noise level of the Slack channel low.

Change-Id: Idda5ba6147cd1a21efb5382580e5c576a4821987